Remove all def files that depend on OpenMPI 4.0.4
Removed all container definition files that depend on OpenMPI 4.0.4 due
to the bug/issue described in commits 210cd3e and 8c59ea2. While OpenMPI
4.0.4 is currently the base OpenMPI version deployed on Expanse @ SDSC,
running Singularity containers with OpenMPI 4.0.5 against OpenMPi 4.0.4
on the host has not been a problem. Moreover, we will plan to update the
base OpenMPI version on Expanse to OpenMPI 4.0.5 on next update of Spack
instances (in the next month or two).