Add Singularity.spark-2.3.1-hadoop-2.7-ubuntu-18.04
This is an updated definition file for Spark based on the example
definition file provided by Sylabs [1] a few years ago. It has been
tested on Comet and is expected to serve as an initial testbed container
to explore multi-node Spark cluster configruations on Expanse. e.g., the
upcoming CIML workshop plans to use PySpark during a number of their
training sessions.

[1] https://github.com/sylabs/examples/tree/master/machinelearning/spark