Include apt-get install yum-utils during installation for Ubuntu
This is required to allow yum-based bootstraping of definition files.
e.g., CentOS-based containers. Note, however, there are currently no deb
packages for yum or yum-utils in the standard Ubuntu 20.04 LTS (focal)
repositories. An alternative approach to installing either yum and/or
dnf will be required in the future.

Also fixed a typo of a missing colon on a call to the log::error
function when naked-singularity.sh checks for a pre-existing
installation of Singularity.