Add CentOS 7.7 and CentOS 7.7 + OpenMPI 3.1.6 def files
An Expanse user requires a much older GCC compiler for their code, but
also still wants to run the code across multiple nodes with full
Infiniband support.

Please note that CentOS 7.7 has been deprecated and should only be used
in limited circumstances when necessary. In this case, Mellanox does not
provide an OFED version matching the one we have on Expanse for the
latest versions of CentOS 7.