Add CentOS 7.7 + CUDA 11.0 + OpenMPI 3.1.6/4.0.5 def files
As mentioned in a recent commit, an Expanse user required a much older
GCC compiler for their code, but also still wanted to run the code
across multiple nodes with full Infiniband support. This code also
required a CUDA-enabled version of OpenMPI. As such, these additional
definition files were created to support their CUDA requirement.
Infiniband support was tested with the OMB containers also included here.

Again, please note that CentOS 7.7 has been deprecated and should only
be used in limited circumstances when necessary. In this case, Mellanox
does not provide an OFED version matching the one we have on Expanse for
the latest versions of CentOS 7.