Update Ubuntu + MLNX_OFED + OpenMPI def files; explicit ./configure
Updated the definition files for the base Ubuntu 18.04 + Mellanox
OFED 4.7 + OpenMPI 3.1.6 and 4.0.5 Singularity containers to have more
extensive and explicit ./configure commands for each OpenMPI build. This
is to to help investigate higher than expected latencies observed with
collective OMB benchmarks when using UCX-only OpenMPI 4.0.5 on Expanse.
Note, however, these high latencies are observed for non-containerized
builds of OMB + OpenMPI 4.0.5 as well and are not specific to the
containerzed builds themselves. The issue appears to be UCX-specific at
this time, but it is still under investigation.