Update Singularity.ubuntu-18.04-cuda-10.2-mlnx-ofed-4.6-1.0.1.1-openmpi-3.1.4
Including its base and intermediate layers. In particular, please note
that MLNX-OFED and OpenMPI layers no longer update all of the software
within the container near the beginning of the %post section of the
definition files. This led to an unintended update of the CUDA version
and its libraries in certain cases --- e.g., when the minor version
could be updated; I don't believe major versions were updated
automatically.