Update Singularity.ubuntu-18.04-cuda-11.2-mlnx-ofed-4.7-3.2.9.0-openmpi-4.0.5
Including its base and intermediate layers. Again, the MLNX-OFED and
OpenMPI layers no longer update all of the software within the container
by default near the beginning of the %post section. It was observed with
this set of containers in particular that an unintended update of the
minor version of the NCCL libraries, which caused a runtime problem when
subsequentely used for Horovod + TensorFlow. Apparentely there was a
runtime API change in CUDA 11.3.