Add Singularity.pytorch-1.8.2-ubuntu-18.04-cuda-11.2-mlnx-ofed-4.7-3.2.9.0-openmpi-4.0.5
Updated PyTorch to v1.8.2, which is expected to be first version
available on Voyager. The container now includes PyTorch Lightning,
Horovod, and ONNX. Note the small fix/patch applied to PyTorch
Lightning in order to support TorchText integration.

Also corrected source build instructions of Horovod for TensorFlow in
the latest definition file.