Add Singularity.ubuntu-18.04-cuda-11.2-mlnx-ofed-4.7-3.2.9.0-openmpi-3.1.6
Currently working on an updated version of the DeepBench container in
the naked-singularity archive. DeepBench's OMB benchmarks require the
C++ language bindings for MPI, which were depreacted from the MPI
standard and removed from OpenMPI starting with the OpenMPI v4.x.x
series.

The DeepBench container will be used to help prepare acceptance tests
for Voyager on Expanse. Several of the proposed microbenchmark
acceptance tests for Voyager are based on the DeepBench GEMM, CONV, and
ALL-REDUCE benchmarks.