Fix "Cannot autolaunch D-Bus without X11 $DISPLAY" error
While working through the on-going, 'comprehensive' update of the
naked-singularity project's definition files to support Expanse's recent
operating system and hardware driver updates, we had to downgrade the
Singularity container build host on Jetstream from Ubuntu 20.04 LTS to
Ubuntu 18.04 LTS in order to support builds of CentOS-based containers.
Again, this is due to the yum-utils deb package not being available in
Ubuntu 20.04. This downgrade occured midway through this update of the
project's definition files and containers. The following build-related
error when bootstrapping a new container build (fenics) via oras from
the GitHub Container Registry was observed.

    INFO:    Starting build...
    FATAL:   While performing build: conveyor failed to get: while fetching library image: failed to get checksum for oras://ghcr.io/mkandes/naked-singularity:ubuntu-18.04-mlnx-ofed-4.9-4.1.7.0-openmpi-3.1.6: while resolving reference: error getting credentials - err: exit status 1, out: `Cannot autolaunch D-Bus without X11 $DISPLAY`

The issue was resolved upon installing the following packages, which are
now included in the naked-singularity.sh install command.

    apt-get -y install pass
    apt-get -y install gnupg2

It's quite likely these packages are simply default dependencies for
other packages in the Singularity installation process under Ubuntu
20.04 LTS, but must be explicitly installed now for Ubuntu 18.04 LTS. At
least, that would seem like the most likely explanation.