Add Singularity.ubuntu-22.04-x86_64
This base Ubuntu 22.04 LTS Singularity container definition file begins
the final reorganization of the naked-singularity project and its
collection of containers. With the increasing use of Docker containers
for other projects at SDSC, in addition to other types of package
software environments, such as conda-packed environments, I plan to
migrate and merge much this software environment work into a new
umbrella project during the coming year.

In anticipation of the EOL for Ubuntu 20.04 LTS in 2025, I also plan to
remove all containers with this as the base operating system during this
transition --- like the last commit on Ubuntu 18.04 LTS. However, when
possible, I also plan to create a new set of containers based on Rocky
Linux 8 to provide for another option for another long-term support
operating system again.

Much of the focus during this transition towards archiving the
naked-singularity project will focus on updating containers that are
important for certain benchmarks that will be used in SDSC projects in
the near-term.

All other application containers that do not recieve a final update
during this transition will be moved to and reogranized under the
existing archive directory within the repository for future reference.