Add Singularity.ubuntu-22.04-mlnx-ofed-5.8-2.0.3.0-openmpi-4.1.5a1-x86_64
Please note that this definition file is identical to the updated
version of Singularity.ubuntu-22.04-mlnx-ofed-5.8-2.0.3.0 also contained
within this commit, except for the difference in the metadata contained
within the %labels section. This variant of the definition file is
intended to explicitly mark the use of the OpenMPI version installed as
part of the MLNX_OFED distribution. All other MPI containers built from
one of these base MLNX_OFED containers must now explicitly override the
use of this native OpenMPI installation via modification to the PATH
and LD_LIBRARY_PATH environment variables.