Archive all Singularity.* (and Dockerfile.*) not in use
This commit archives all existing Singularity.* (and Dockerfile.*) not
currently in active use on Expanse or Voyager at SDSC. The cleanup of
the naked-singularity project repository and its container definition
files is part of a final reorganization as we tranisition to a new
ubrella project at SDSC this year, which will track not only Singularity.*
(and Dockerfile.*), but also conda- and spack-based environments in use
on our systems as well. Once this draft reorganization is complete, which
began at [1] and will serve as the base of the new ubrella project, the
naked-singularity project will be permanently archived on GitHub in
read-only mode.

[1]

https://github.com/mkandes/naked-singularity/commit/a87ab4ed04342ecbaea26cffb5a70452660ca990